fix(lint): resolve all SwiftLint CI violations for PR #460
- AnalyticsLogger: remove redundant LaunchType raw values, extract
  AppLaunchReadinessPayload struct to reduce appLaunchReadiness to 1
  associated value (fixes enum_case_associated_values_count), rename
  binding 'p' -> 'payload' (fixes identifier_name), fix line_length
- AppCoordinator: extract AppGroupIPCProviding protocol to separate
  file (AppGroupIPCProviding.swift), extract ReminderScheduling
  conformance (AppCoordinator+ReminderScheduling.swift), and move
  Fallback Timer Shims extension (AppCoordinator+Helpers.swift) to
  resolve file_length (638->~325 lines) and type_body_length
- AnalyticsLoggerTests, AnalyticsEventTests: update 5 appLaunchReadiness
  callsites to use AppLaunchReadinessPayload(.init(...))
- SettingsFlowTests: rename masterToggle -> globalToggle (fixes
  inclusive_language), shorten NSPredicate string (fixes line_length)
- SettingsViewModelExtendedTests: replace 5 force-unwraps with
  optional chaining (fixes force_unwrapping)
- ComponentsExtendedTests: split long assertion line (fixes
  line_length), fix redundant_discardable_let
- ViewBodyCoverageTests: remove trailing comma (fixes trailing_comma)
- AppCoordinatorNotificationFallbackTests: split long awaitCondition
  call (fixes line_length)
- AppCoordinatorTests: add makeTrackedCoordinator() helper, refactor
  3 handleNotification tests to use it (fixes file_length 606->~599)
- AppGroupIPCStoreTests: fix comma spacing (4 instances), rename
  'i' -> 'idx' (fixes identifier_name), compact loops (fixes
  type_body_length 406->~398)
